+++ packages/fuchsia_ctl/CHANGELOG.md
@@ -1,14 +1,16 @@
 # CHANGELOG
 
+## 0.0.22
 
+- Added `emu` tool for spawning an emulator instance given a Fuchsia QEMU build,
+  the Fuchsia SDK, and an Android emulator executable.
+- Fixed homepage link.
 
 ## 0.0.18 - 0.0.21
 
 - Add retries to paving operations.
 - Add timeouts to paving and ssh operations.
 - Add arguments parameter to test command.
-
-
 
 ## 0.0.17